Move hashpash to parent method
