+++ demo/src/main/java/urlshortener/demo/domain/URIBase.java
@@ -1,0 +1,46 @@
+package urlshortener.demo.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.annotations.ApiModelProperty;
+
+import javax.validation.constraints.NotNull;
+import java.util.Objects;
+
+public class URIBase {
+
+    @JsonProperty("hashpass")
+    private String hashpass = null;
+
+    public URIBase hashpass(String hashpass) {
+        this.hashpass = hashpass;
+        return this;
+    }
+
+    /**
+     * Get hashpass
+     * @return hashpass
+     **/
+    @ApiModelProperty(example = "adsgv674fjhag", required = true, value = "")
+    @NotNull
+
+
+    public String getHashpass() {
+        return hashpass;
+    }
+
+    public void setHashpass(String hashpass) {
+        this.hashpass = hashpass;
+    }
+
+    @Override
+    public boolean equals(java.lang.Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        URIBase urIItem = (URIItem) o;
+        return Objects.equals(this.hashpass, urIItem.hashpass);
+    }
+}
