+++ demo/src/main/java/urlshortener/demo/domain/URIItem.java
@@ -13,15 +13,12 @@ import java.util.Objects;
 @Validated
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2018-11-26T14:20:22.002Z[GMT]")
 
-public class URIItem   {
+public class URIItem extends URIBase {
   @JsonProperty("id")
   private String id = null;
 
   @JsonProperty("redirection")
   private String redirection = null;
-
-  @JsonProperty("hashpass")
-  private String hashpass = null;
 
   public URIItem id(String id) {
     this.id = id;
@@ -65,27 +62,6 @@ public class URIItem   {
     this.redirection = redirection;
   }
 
-  public URIItem hashpass(String hashpass) {
-    this.hashpass = hashpass;
-    return this;
-  }
-
-  /**
-   * Get hashpass
-   * @return hashpass
-  **/
-  @ApiModelProperty(example = "adsgv674fjhag", required = true, value = "")
-  @NotNull
-
-
-  public String getHashpass() {
-    return hashpass;
-  }
-
-  public void setHashpass(String hashpass) {
-    this.hashpass = hashpass;
-  }
-
 
   @Override
   public boolean equals(java.lang.Object o) {
@@ -96,14 +72,13 @@ public class URIItem   {
       return false;
     }
     URIItem urIItem = (URIItem) o;
-    return Objects.equals(this.id, urIItem.id) &&
-        Objects.equals(this.redirection, urIItem.redirection) &&
-        Objects.equals(this.hashpass, urIItem.hashpass);
+    return super.equals(o) && Objects.equals(this.id, urIItem.id) &&
+        Objects.equals(this.redirection, urIItem.redirection);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(id, redirection, hashpass);
+    return Objects.hash(id, redirection, getHashpass());
   }
 
   @Override
@@ -113,7 +88,7 @@ public class URIItem   {
     
     sb.append("    id: ").append(toIndentedString(id)).append("\n");
     sb.append("    redirection: ").append(toIndentedString(redirection)).append("\n");
-    sb.append("    hashpass: ").append(toIndentedString(hashpass)).append("\n");
+    sb.append("    hashpass: ").append(toIndentedString(getHashpass())).append("\n");
     sb.append("}");
     return sb.toString();
   }
